requirements.txt
# Core Flask packages
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0

# Database packages
PyMySQL==1.1.0
mysql-connector-python==8.1.0

# Security packages
Werkzeug==2.3.6
bcrypt==4.0.1

# Additional utilities
python-dotenv==1.0.0
Pillow>=9.0.0